Extracted names builders into their own namespace
