+++ Service/NamesBuilders/IArmyNamesBuilder.cs
@@ -1,0 +1,9 @@
+using StellarisNameListGenerator.Models;
+
+namespace StellarisNameListGenerator.Service.NamesBuilders
+{
+    public interface IArmyNamesBuilder
+    {
+        string Build(NameList nameList);
+    }
+}
