+++ Service/NamesBuilders/IShipClassNamesBuilder.cs
@@ -1,0 +1,9 @@
+using StellarisNameListGenerator.Models;
+
+namespace StellarisNameListGenerator.Service.NamesBuilders
+{
+    public interface IShipClassNamesBuilder
+    {
+        string Build(NameList nameList);
+    }
+}
